+++ ios/StripeSdk.swift
@@ -48,6 +48,17 @@ class StripeSdk: RCTEventEmitter, STPBankSelectionViewControllerDelegate, UIAdap
     var applePayShippingAddressErrors: [Error]? = nil
     var applePayCouponCodeErrors: [Error]? = nil
     
+    var customerSheetConfiguration = CustomerSheet.Configuration()
+    var customerSheet: CustomerSheet? = nil
+    var customerAdapter: StripeCustomerAdapter? = nil
+    var customerSheetViewController: UIViewController?
+    var fetchPaymentMethodsCallback: (([STPPaymentMethod]) -> Void)? = nil
+    var attachPaymentMethodCallback: (() -> Void)? = nil
+    var detachPaymentMethodCallback: (() -> Void)? = nil
+    var setSelectedPaymentOptionCallback: (() -> Void)? = nil
+    var fetchSelectedPaymentOptionCallback: ((CustomerPaymentOption?) -> Void)? = nil
+    var setupIntentClientSecretForCustomerAttachCallback: ((String) -> Void)? = nil
+    
     var hasEventListeners = false
     override func startObserving() {
         hasEventListeners = true
@@ -57,7 +68,9 @@ class StripeSdk: RCTEventEmitter, STPBankSelectionViewControllerDelegate, UIAdap
     }
     
     override func supportedEvents() -> [String]! {
-        return ["onOrderTrackingCallback", "onConfirmHandlerCallback"]
+        return ["onOrderTrackingCallback", "onConfirmHandlerCallback", "onCustomerAdapterFetchPaymentMethodsCallback", "onCustomerAdapterAttachPaymentMethodCallback",
+        "onCustomerAdapterDetachPaymentMethodCallback", "onCustomerAdapterSetSelectedPaymentOptionCallback", "onCustomerAdapterFetchSelectedPaymentOptionCallback",
+        "onCustomerAdapterSetupIntentClientSecretForCustomerAttachCallback"]
     }
 
     @objc override static func requiresMainQueueSetup() -> Bool {